refactor: split poll_stages.rs test module into poll_stages_tests.rs
poll_stages.rs was 3,527 LOC, ~2,552 of which was a single inline
#[cfg(test)] mod tests block at the tail. Relocate that block into a
sibling poll_stages_tests.rs, loaded via #[cfg(test)] #[path =
"poll_stages_tests.rs"] mod tests; — the same pure-code-motion test-split
pattern used across the userspace-dp modularity work (#4409 nat tests,
etc.). This drops the production file to ~975 LOC, making the poll-stage
production code reviewable without scrolling past 2.5k lines of tests
(part of the #4404/#4408/#4421 modularity backlog).

Pure code motion, verified three ways:
- The test-module BODY is byte-identical: the relocated file dedented one
  indent level, then re-indented and re-wrapped in `mod tests { ... }`,
  reproduces the original block exactly (checked line-for-line vs
  origin/master).
- The production prefix (lines 1-972) is byte-identical to origin/master
  (unified diff empty).
- Test count is preserved: 26 #[test]/#[tokio::test] attributes before and
  after. The module keeps `use super::*` so it still resolves into
  poll_stages.rs internals; parent is still `afxdp::poll_stages` (the
  #[path] child keeps its module parent), so super:: paths are unchanged.

Validation: cargo build clean (the #[path] split + use-clauses resolve);
FULL cargo test --release to follow before merge. No production behavior
change. Refs #4408 #4421.
